msm: ep_pcie: Update phy sequence in D3 hot
Some customer platforms do not support D3 cold. To be able
to support L1ss on those platforms where the link gets
initialized with the PBL phy sequence - which may not be
optimal for L1ss - the latest phy sequence is updated
during handling of the first D3 hot interrupt.

Change-Id: If47d6b2c4ddab6b26f7ad1d98bd0f40d054d4d9f
Signed-off-by: Siva Kumar Akkireddi <sivaa@codeaurora.org>